+++ lib/src/media_stream_impl.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:js_interop';
+
 import 'package:web/web.dart' as web;
 import 'package:webrtc_interface/webrtc_interface.dart';