+++ lib/src/mediadevices_impl.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:js' as js;
 import 'dart:js_interop';
 import 'dart:js_util' as jsutil;
+
 import 'package:web/web.dart' as web;
 import 'package:webrtc_interface/webrtc_interface.dart';
 
@@ -13,15 +14,41 @@ class MediaDevicesWeb extends MediaDevices {
   Future<MediaStream> getUserMedia(
       Map<String, dynamic> mediaConstraints) async {
     try {
-      if (!isMobile) {
-        if (mediaConstraints['video'] is Map &&
-            mediaConstraints['video']['facingMode'] != null) {
-          mediaConstraints['video'].remove('facingMode');
+      try {
+        if (!isMobile) {
+          if (mediaConstraints['video'] is Map &&
+              mediaConstraints['video']['facingMode'] != null) {
+            mediaConstraints['video'].remove('facingMode');
+          }
         }
+        mediaConstraints.putIfAbsent('video', () => false);
+        mediaConstraints.putIfAbsent('audio', () => false);
+      } catch (e) {
+        print(
+            '[getUserMedia] failed to remove facingMode from mediaConstraints');
       }
+      try {
+        if (mediaConstraints['audio'] is Map<String, dynamic> &&
+            Map.from(mediaConstraints['audio']).containsKey('optional') &&
+            mediaConstraints['audio']['optional']
+                is List<Map<String, dynamic>>) {
+          List<Map<String, dynamic>> optionalValues =
+              mediaConstraints['audio']['optional'];
+          final audioMap = <String, dynamic>{};
 
-      mediaConstraints.putIfAbsent('video', () => false);
-      mediaConstraints.putIfAbsent('audio', () => false);
+          optionalValues.forEach((option) {
+            option.forEach((key, value) {
+              audioMap[key] = value;
+            });
+          });
+
+          mediaConstraints['audio'].remove('optional');
+          mediaConstraints['audio'].addAll(audioMap);
+        }
+      } catch (e, s) {
+        print(
+            '[getUserMedia] failed to translate optional audio constraints, $e, $s');
+      }
 
       final mediaDevices = web.window.navigator.mediaDevices;