+++ lib/src/utils.dart
@@ -1,4 +1,6 @@
+import 'dart:js_util' as jsutil;
 import 'dart:math';
+
 import 'package:web/web.dart' as web;
 
 bool get isMobile {